third_party/WebKit/LayoutTests/compositing/overflow/scrolls-with-respect-to-transform-expected.txt

This test ensures that Layer::scrollsWithRespectTo always works even with transformed fixed-position layers.

fixed does not scroll with respect to fixed
fixed scrolls with respect to container
fixed scrolls with respect to overflow-child
fixed scrolls with respect to abs-descendant
fixed scrolls with respect to fixed-descendant
fixed scrolls with respect to viewportFiller
fixed scrolls with respect to root
container scrolls with respect to fixed
container does not scroll with respect to container
container scrolls with respect to overflow-child
container scrolls with respect to abs-descendant
container scrolls with respect to fixed-descendant
container does not scroll with respect to viewportFiller
container does not scroll with respect to root
overflow-child scrolls with respect to fixed
overflow-child scrolls with respect to container
overflow-child does not scroll with respect to overflow-child
overflow-child does not scroll with respect to abs-descendant
overflow-child does not scroll with respect to fixed-descendant
overflow-child scrolls with respect to viewportFiller
overflow-child scrolls with respect to root
abs-descendant scrolls with respect to fixed
abs-descendant scrolls with respect to container
abs-descendant does not scroll with respect to overflow-child
abs-descendant does not scroll with respect to abs-descendant
abs-descendant does not scroll with respect to fixed-descendant
abs-descendant scrolls with respect to viewportFiller
abs-descendant scrolls with respect to root
fixed-descendant scrolls with respect to fixed
fixed-descendant scrolls with respect to container
fixed-descendant does not scroll with respect to overflow-child
fixed-descendant does not scroll with respect to abs-descendant
fixed-descendant does not scroll with respect to fixed-descendant
fixed-descendant scrolls with respect to viewportFiller
fixed-descendant scrolls with respect to root
viewportFiller scrolls with respect to fixed
viewportFiller does not scroll with respect to container
viewportFiller scrolls with respect to overflow-child
viewportFiller scrolls with respect to abs-descendant
viewportFiller scrolls with respect to fixed-descendant
viewportFiller does not scroll with respect to viewportFiller
viewportFiller does not scroll with respect to root
root scrolls with respect to fixed
root does not scroll with respect to container
root scrolls with respect to overflow-child
root scrolls with respect to abs-descendant
root scrolls with respect to fixed-descendant
root does not scroll with respect to viewportFiller
root does not scroll with respect to root
